Add files via upload
This is my progress on a weekly game jam by WGJ with the Theme "Tennis." The game is intended to be similar to a rhythm game. Currently, there are 3 blocks that emit colors in a straight line, and the player is a paddle that can hit the colors to increment a score.